allow more cases to use the let clause; fixes #3563
